Change listener to be floats

+++ src/entity/engineComponents.h
@@ -119,8 +119,8 @@ struct Workspace
 
 struct Listener
 {
-  int x = 0;
-  int y = 0;
-  int z = 0;
-  int gain = 1;
+  float x = 0;
+  float y = 0;
+  float z = 0;
+  float gain = 1;
 };
